Remove boilerplate, and format
