+++ fmm/src/lib.rs
@@ -1,5 +1,6 @@
 //! Fast Solver FMM library
 #![cfg_attr(feature = "strict", deny(warnings))]
 pub mod constants;
+pub mod fmm;
 pub mod laplace;
 pub mod linalg;
