+++ tree/src/types/node.rs
@@ -1,5 +1,7 @@
+/// Types for handling data on tree nodes.
 use crate::types::{morton::MortonKey, point::Points};
 
+/// A LeafNode contains Points which define the discretisation, and is indexed by a MortonKey.
 #[derive(Debug, Clone, Default)]
 pub struct LeafNode {
     pub key: MortonKey,
@@ -8,6 +10,8 @@ pub struct LeafNode {
 
 pub type LeafNodes = Vec<LeafNode>;
 
+/// A Node contains a generic data vector which a user can create views for specific to
+/// their application.
 #[derive(Debug, Clone, Default)]
 pub struct Node {
     pub key: MortonKey,
@@ -16,6 +20,7 @@ pub struct Node {
 
 pub type Nodes = Vec<Node>;
 
+/// NodeData provides a wrapper for raw data stored in each node.
 #[derive(Debug, Clone, Default)]
 pub struct NodeData {
     pub field_size: Vec<usize>,
